moved concerns for SC2 and WoW to their own directories
