+++ sc2/profile.go
@@ -1,0 +1,5 @@
+package sc2
+
+import(
+	"github.com/mitchellh/go-bnet"
+)
